requirements.txt
pillow
python-dotenv
moviepy
google-genai
elevenlabs